updating

+++ scripts/lib/recursive-copy.js
@@ -3,9 +3,11 @@
 const path = require('path')
 const fs = require('fs')
 
-const recursiveCopy = (from, to) => {
+const recursiveCopy = (from, to, overwrite = true) => {
   if (!fs.statSync(from).isDirectory()) {
-    fs.copyFileSync(from, to)
+    if (!fs.existsSync(to) || overwrite) {
+      fs.copyFileSync(from, to)
+    }
   } else {
     if (!fs.existsSync(to)) {
       fs.mkdirSync(to)
